Allow simple running of FC
The current implementation expects a full configuration of the output format to be passed in when processing some Fractional Cover. This is unnecessary for a lot of use cases, as we can simple define a default output structure.

This could potentially also simplify the large scale processing setup, but would have to combine extra metadata with the low level datatype and variable name options.